Added ToolkitFunctional.js
- consolidates all Toolkit methods into one test file
- remove deprecated files
