--- test/functional/iDataQueueFunctional.js
@@ -1,129 +1,0 @@
-// Copyright (c) International Business Machines Corp. 2019
-// All Rights Reserved
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
-// associated documentation files (the "Software"), to deal in the Software without restriction,
-// including without limitation the rights to use, copy, modify, merge, publish, distribute,
-// sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in all copies or
-// substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
-// NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-/* eslint-env mocha */
-/* eslint-disable new-cap */
-
-const { expect } = require('chai');
-const { Toolkit } = require('../../lib/itoolkit');
-
-// Set Env variables or set values here.
-const opt = {
-  database: process.env.TKDB || '*LOCAL',
-  username: process.env.TKUSER || '',
-  password: process.env.TKPASS || '',
-  host: process.env.TKHOST || 'localhost',
-  port: process.env.TKPORT || 80,
-  path: process.env.TKPATH || '/cgi-bin/xmlcgi.pgm',
-};
-
-const lib = 'NODETKTEST'; const dqName = 'TESTQ';
-
-const { returnTransports } = require('../../lib/utils');
-
-const transports = returnTransports(opt);
-
-describe('iDataQueue Functional Tests', () => {
-  before('setup library for tests and create DQ', async () => {
-    // eslint-disable-next-line global-require
-    const { DBPool } = require('idb-pconnector');
-
-    const pool = new DBPool({ url: '*LOCAL' }, { incrementSize: 2 });
-
-    const qcmdexec = 'CALL QSYS2.QCMDEXC(?)';
-
-    const createLib = `CRTLIB LIB(${lib}) TYPE(*TEST) TEXT('Used to test Node.js toolkit')`;
-
-    const createDQ = `CRTDTAQ DTAQ(${lib}/${dqName}) MAXLEN(100) AUT(*EXCLUDE) TEXT('TEST DQ FOR NODE TOOLKIT TESTS')`;
-
-    const findLib = 'SELECT SCHEMA_NAME FROM qsys2.sysschemas WHERE SCHEMA_NAME = \'NODETKTEST\'';
-
-    const findDQ = 'SELECT OBJLONGNAME FROM TABLE (QSYS2.OBJECT_STATISTICS(\'NODETKTEST\', \'*DTAQ\')) AS X';
-
-    const libResult = await pool.runSql(findLib);
-
-    const dqResult = await pool.runSql(findDQ);
-
-    if (!libResult.length) {
-      await pool.prepareExecute(qcmdexec, [createLib]).catch((error) => {
-        // eslint-disable-next-line no-console
-        console.log('Unable to Create Lib!');
-        throw error;
-      });
-      // eslint-disable-next-line no-console
-      console.log('CREATED LIB!');
-    }
-    if (!dqResult.length) {
-      await pool.prepareExecute(qcmdexec, [createDQ]).catch((error) => {
-        // eslint-disable-next-line no-console
-        console.log('Unable to Create DQ!');
-        throw error;
-      });
-      // eslint-disable-next-line no-console
-      console.log('CREATED DQ!');
-    }
-  });
-
-  describe('sendToDataQueue', () => {
-    transports.forEach((transport) => {
-      it(`sends data to specified DQ using ${transport.name} transport`, (done) => {
-        const connection = transport.me;
-
-        const toolkit = new Toolkit(connection);
-
-        toolkit.sendToDataQueue(dqName, lib, 'Hello from DQ!', (error, output) => {
-          expect(error).to.equal(null);
-          expect(output).to.equal(true);
-          done();
-        });
-      });
-    });
-  });
-
-  describe('receiveFromDataQueue', () => {
-    transports.forEach((transport) => {
-      it(`receives data from specfied DQ using ${transport.name} transport`, (done) => {
-        const connection = transport.me;
-
-        const toolkit = new Toolkit(connection);
-
-        toolkit.receiveFromDataQueue(dqName, lib, 100, (error, output) => {
-          expect(error).to.equal(null);
-          expect(output).to.be.a('string').and.to.equal('Hello from DQ!');
-          done();
-        });
-      });
-    });
-  });
-
-  describe('clearDataQueue', () => {
-    transports.forEach((transport) => {
-      it(`clears the specifed DQ using ${transport.name} transport`, (done) => {
-        const connection = transport.me;
-
-        const toolkit = new Toolkit(connection);
-
-        toolkit.clearDataQueue(dqName, lib, (error, output) => {
-          expect(error).to.equal(null);
-          expect(output).to.equal(true);
-          done();
-        });
-      });
-    });
-  });
-});
